+++ tests/Feature/Api/Files/ListFilesTest.php
@@ -2,6 +2,7 @@
 
 namespace Tests\Feature\Api\Files;
 
+use Illuminate\Http\Response;
 use Tests\RequiresAuth;
 use Tests\TestCase;
 
@@ -29,6 +30,25 @@ class ListFilesTest extends TestCase
             'filename' => 'hello.md',
             'filepath' => resource_path('test-skel/mixed'),
             'mimetype' => 'text/plain',
+        ]);
+    }
+
+    /** @test */
+    public function nonexistant_paths_return_error_with_filepath(): void
+    {
+        $response = $this->authed()->getJson($this->endpoint([
+            'path' => '/tmp/nothere',
+        ]), [
+            'contents' => 'invalid file',
+        ]);
+
+        $response->assertStatus(Response::HTTP_NOT_FOUND);
+        $response->assertJson([
+            'filepath' => '/tmp/nothere',
+            'error' => [
+                'code' => 404,
+                'msg' => "This directory doesn't exist.",
+            ],
         ]);
     }
 }